Improved performance by using output parameters.
By passing a std::vector<TraceRes>* and returning void instead of just
returning a std::vector<TraceRes>, the implementation can reuse the same
vector as the output parameter for several traces. This reduces the
number of allocations that are necessary. Furthermore, the
implementation for a trace can do intermediate work inside the output
vector.

Overall, this makes the raytracer more than twice as fast.