Added diffraction to glass.
This is slow. It causes 3x as many rays with every glass boundary which
is hit. It is not as trivial to add this feature without such blow-up,
but it can be achieved by making the Trace function aware of what colour
multipliers are being applied to the result. This could even be used
instead of the "significance" parameter.